refactor(event coordinator): remove event coordinator
I had to reflect a little on this decision, but it's quite simple at the
end of the day. Right now, the event coorindator allows a link for a
domain found on one process to be added to a frontier managed by a
different process. The idea is that we don't want to miss links.

But if you give this plan thought, it's kind of silly. What are the
chances that a link to domain X is found on an external domain that does not also
exist on a page on domain X? Probably quite low. The technical
complexity introduced by IPC adds little value.