test(robots): add tests to check error status codes
I also needed to require the user to create an instance of the robot's
isAllowed function. We needed to pass in a copy of the cache to prevent
tests from interacting with eachother (previously the cache was created
at the module level). Also, I learned a lot about why singletons are an
anti-pattern. I need to re-do the logging module to make parts of this
code testable.