fix(robots parser): add timeout
Timeouts are absolutely critical. It turns out that the vast majority of
sites don't response to robotsTxt requests (at least from my
useragent/setup). We don't want to wait seconds on every single one of
these when, in general, for sites that do response to robotsTxt, the
request is fast